requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.28.2
python-dotenv==0.21.0
huggingface_hub==0.13.3
langchain==0.0.153
PyPDF2==3.0.1
transformers==4.28.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
